[feat-7-py]: Create file path inside main function
Adjusted the Python code solution for day 7 so that the file path
is created inside the main function and passed to the file reading
function.

Additionally, I'm trying a new commit convention where I include the
day number and optionally the programming language alongside
the commit type so I don't have to continually specify the day and
language repeatedly throughout the commit message. I'll try this
with my other AoC repositories.